+++ src/main/java/fr/mrjuju0319/pvoteparty/vote/VoteStorage.java
@@ -37,6 +37,8 @@ public interface VoteStorage {
 
     void resetAllPalliers(String playerName);
 
+    void resetPalliersForAllPlayers(String pallierOrAll);
+
     void upsertSharedConfig(VoteConfig config);
 
     SharedConfig loadSharedConfig();